Add security enhancements, streaming fixes, and advanced UAC bypass
